+++ backend/dps_training_k/deployment/nginx/nginx_deploy_prod.conf
@@ -12,13 +12,24 @@ server {
     ssl_certificate /etc/nginx/ssl/fullchain.pem;
     ssl_certificate_key /etc/nginx/ssl/privkey.pem;
 
+    # Route to React app for landing page and other routes
     location / {
+        proxy_pass http://frontend-react:3000;
+        proxy_set_header Host $host;
+        proxy_set_header X-Real-IP $remote_addr;
+        proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+        proxy_set_header X-Forwarded-Proto $scheme;
+    }
+
+    # Route to Vue app for simulation
+    location /simulation {
         proxy_pass http://frontend:5173;
         proxy_set_header Host $host;
         proxy_set_header X-Real-IP $remote_addr;
         proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
         proxy_set_header X-Forwarded-Proto $scheme;
     }
+
     location /api {
         proxy_pass http://django:8000/api;
         proxy_set_header Host $host;
