+++ frontend-react/src/pages/LoginPage.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+const LoginPage: React.FC = () => {
+  return (
+    <div className="page-container">
+      <h1>Login</h1>
+      <p>Login page - to be implemented</p>
+    </div>
+  );
+};
+
+export default LoginPage;
