+++ frontend-react/src/pages/PatientPage.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+const PatientPage: React.FC = () => {
+  return (
+    <div className="page-container">
+      <h1>Patient</h1>
+      <p>Patient page - to be implemented</p>
+    </div>
+  );
+};
+
+export default PatientPage;
