+++ frontend-react/src/pages/TrainerPage.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+const TrainerPage: React.FC = () => {
+  return (
+    <div className="page-container">
+      <h1>Trainer</h1>
+      <p>Trainer page - to be implemented</p>
+    </div>
+  );
+};
+
+export default TrainerPage;
